chore: replace newsletter confirm route with server lib fn
